+++ src/main/java/org/choongang/Application.java
@@ -1,6 +1,6 @@
 package org.choongang;
 
-import org.choongang.global.main.MainRouter;
+import org.choongang.main.MainRouter;
 
 public class Application {
     public static void main(String[] args) {
