--- src/main/java/org/choongang/constants/Menu.java
@@ -1,9 +1,0 @@
-package org.choongang.constants;
-
-public enum Menu {
-    MAIN,//메인화면
-    JOIN,//회원가입
-    LOGIN,//로그인
-    GAME,//게임시작
-    RESULT//결과
-}
