--- src/main/java/org/choongang/global/main/MainController.java
@@ -1,10 +1,0 @@
-package org.choongang.global.main;
-
-import org.choongang.global.AbstractController;
-
-public class MainController extends AbstractController {
-    @Override
-    public void show() {
-        System.out.println("메인메뉴");
-    }
-}
