+++ src/main/java/org/choongang/member/controller/JoinController.java
@@ -1,4 +1,4 @@
-package org.choongang.global.member.controller;
+package org.choongang.member.controller;
 
 import org.choongang.global.AbstractController;
